+++ apiphany-core/src/main/java/org/apiphany/security/keys/RSAKeys.java
@@ -14,7 +14,7 @@ import org.morphix.reflection.Constructors;
  * RSA key management utility methods.
  * <p>
  * This class provides methods to load RSA public and private keys from PEM-formatted files. No fancy dependencies are
- * used for PEM parsing; the implementation relies on standard Java libraries. Also no fancy validation is performed on
+ * used for PEM parsing; the implementation relies on standard Java libraries, also no fancy validation is performed on
  * the keys, assuming the files are well-formed PEM files.
  * <p>
  * If the files cannot be read or the keys cannot be parsed, a {@link SecurityException} is thrown with details about