+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 ## Release Notes
 
+`1.1.2`
+
+- Added `ClientProperties.setBaseUrl` and `ClientProperties.setBaseUrl` to be able to configure the base URL per exchange client.
+- Renamed `ServerFinished` to `ServerFinishedEncrypted` to better describe the TLS object.
+- Added `ClientFinishedEncrypted` for TLS client finished encrypted handshake.
+- Renamed `Record.getHandshake` to `getHandshakeBody` where the return type is `TLSHandshakeBody`.
+- Renamed `Record.hasHandshake` to `hasHandshakeBody`.
+- Renamed `Record.hasNoHandshake` to `hasNoHandshakeBody`.
+- Renamed `BulkCipher.algorithm` to `jcaKeyAlgorithm` to better emphasize its purpose.
+- Added `BulkCipher.aeadCiphers` static method to return all AEAD ciphers.
+- Added `BulkCipher.blockCiphers` static method to return all BLOCK ciphers.
+- Added `BulkCipher.streamCiphers` static method to return all STREAM ciphers.
+- Added `BulkCipher.encryptingCiphers` static method to return all encrypting ciphers.
+- Added `BulkCipher.encryptingCiphers(CipherType)` static method to return all encrypting ciphers filtered by the given cipher type.
+
+---
+
 `1.1.1`
 
 - Added Maven plugin management to parent `pom.xml` with default configurations - no executions in parent `pluginManagement`.